@@ -1,9 +1,8 @@
 import { Injectable } from '@angular/core';
 import { Actions, Effect, ofType } from '@ngrx/effects';
 import { Action } from '@ngrx/store';
-import { of } from 'rxjs/observable/of';
+import { of ,  Observable } from 'rxjs';
 import { catchError, map, switchMap } from 'rxjs/operators';
-import { Observable } from 'rxjs/Rx';
 import { HttpResponse } from '@angular/common/http';
 
 import {
